backend/requirements.txt
asgiref==3.7.2
boto3==1.34.34
botocore==1.34.34
Django==4.2.7
django-autoslug==1.9.9
django-cors-headers==4.3.1
django-phonenumber-field==7.3.0
django-storages==1.14.2
django-tinymce==3.6.1
djangorestframework==3.14.0
drf-yasg==1.21.7
inflection==0.5.1
jmespath==1.0.1
packaging==23.2
phonenumberslite==8.13.28
pillow==10.2.0
psycopg2-binary
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
PyYAML==6.0.1
s3transfer==0.10.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.4
uritemplate==4.1.1
